Preparing for TUN/TAP
