+++ Netch/Forms/Mode/Process.resx
